pyproject.toml
[tool.poetry]
name = "log-my-printer"
version = "0.1.0"
description = ""
authors = ["NyxAether <contact.nyxhemera@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pywin32 = "^306"
requests = "^2.31.0"
pyyaml = "^6.0.1"
appdirs = "^1.4.4"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
black = "^24.4.0"
mypy = "^1.9.0"
pylint = "^3.1.0"
pytest = "^8.1.1"
isort = "^5.13.2"
pydantic = "^2.7.0"
types-pyyaml = "^6.0.12.20240311"
types-requests = "^2.31.0.20240406"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
